requirements.txt
Crawl4AI
langchain
langchain-community
langchain-openai
langchain-pinecone
langchain-text-splitters
langchain-groq
requests
playwright
pinecone
python-dotenv
tiktoken
openai

aiohttp
aiofiles